♻️🎪 Move animation card
