+++ apps/show/src/exhibitors/animations/card-animations-on-stand.tsx
